#45 기존에 흩어져있던 myPage기능들을 하나의 컨트롤러로 통합
